create the url in an other file and fetching on other url also
